Basic proxy support
Added really basic proxy support. It works for the simple use case of
got("reddit.com") which will first make a http request and then be
redirected to https. This also supports requests that are redirected and
over to unsecure locations i.e. they start http and stay http.

Currently requests that require https to begin with are not supported.