Modify error test since we removed setting body to an object
This test flagged setting a body to something other than ReadableStream,
string, Buffer or plain object. Since we removed the json shortcut
you'll have to always stringify yourself which means you can no longer
set the body to a plain object, and so we won't suggest it in our type
error.