enable sending plain string over rest
while working with a private API i need the functionality to send simple strings with content-type='application/json'. This was not possible with got (or i didnt find a correct way to do it). 

I change the "canBodyBeStringified" calculation so that it is also true if the content-type headers are set to application/json.