Update readme titles
Hi! Amazing lib you've here. We're using it in our project.
Here's a suggestion for the doc which is quite hard to read yet mainly because of the titles.
I just have change titles to make them biggers. I didn't always get the natural hierarchy so there's margin to improve. It's a very small proposal, but huge help to new people discovering your extension.
All the best
Jonas